requirements.txt
# Strands SDK and AI
strands-sdk>=0.1.0
anthropic>=0.39.0

# Hedera SDK
hedera-sdk-python>=2.0.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
aiosqlite>=0.20.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Ethereum/Web3 (for ERC-8004)
web3>=6.0.0
eth-account>=0.13.0
py-solc-x>=2.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.6.0
mypy>=1.11.0
